Ui CardItemProduct V2

+++ shared/src/commonMain/kotlin/com/determent/ecombebop_management/shared/home_content/list_product/a_domain/model/ProductItem.kt
@@ -3,7 +3,7 @@ package com.determent.ecombebop_management.shared.home_content.list_product.a_do
 data class ProductItem(
     val mainImage: Byte,
     val price: Int,
-    val discount: Double,
+    val discount: Int,
     val priceCalculated: Int,
     val title: String,
     val brand: String
